refactor(zodSchema): updated generic zod schema to match the new User Model
