+++ src/common/constants/index.ts
@@ -1,23 +1,23 @@
 export enum Role {
-  SuperUser = 'superuser',
-  User = 'user',
-  Guest = 'guest',
+	SuperUser = 'superuser',
+	User = 'user',
+	Guest = 'guest',
 }
 
 export enum Provider {
-  Local = 'local',
-  Google = 'google',
+	Local = 'local',
+	Google = 'google',
 }
 
 export enum IDType {
-  NIN = 'nin',
-  BVN = 'bvn',
-  IntlPassport = 'intl-passport',
+	NIN = 'nin',
+	BVN = 'bvn',
+	IntlPassport = 'intl-passport',
 }
 
 export enum Gender {
-  Male = 'male',
-  Female = 'female',
-  Others = 'others',
-  None = 'none',
+	Male = 'male',
+	Female = 'female',
+	Others = 'others',
+	None = 'none',
 }
