--- src/lib/zodSchema/loginSchema.ts
@@ -1,11 +1,0 @@
-import { z } from 'zod';
-
-export type LoginSchemaType = typeof LoginSchema;
-
-const LoginSchema = z.object({
-  email: z.string().email('Please enter a valid email address!'),
-  password: z.string().min(12, 'Password must be at least 12 characters!'),
-  rememberMe: z.boolean().optional(),
-});
-
-export { LoginSchema };
